+++ src/features/XIT/ACT/EditMaterialGroup.vue
@@ -4,6 +4,7 @@ import SectionHeader from '@src/components/SectionHeader.vue';
 import Active from '@src/components/forms/Active.vue';
 import TextInput from '@src/components/forms/TextInput.vue';
 import Commands from '@src/components/forms/Commands.vue';
+import Passive from '@src/components/forms/Passive.vue';
 import SelectInput from '@src/components/forms/SelectInput.vue';
 import { act } from '@src/features/XIT/ACT/act-registry';
 
@@ -21,12 +22,13 @@ const nameError = ref(false);
 const typeOptions = act.getMaterialGroupTypes();
 const type = ref(group.type);
 
+const shortDescription = computed(() => act.getMaterialGroupInfo(type.value)?.shortDescription);
 const editFormComponent = computed(() => act.getMaterialGroupInfo(type.value)?.editComponent);
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const editForm = useTemplateRef<any>('editForm');
 
 function onSaveClick() {
-  let isValid = editForm.value.validate();
+  let isValid = editForm.value?.validate() ?? true;
   nameError.value = name.value.length === 0;
   isValid &&= !nameError.value;
   if (!isValid) {
@@ -35,7 +37,7 @@ function onSaveClick() {
   for (const key of Object.keys(group)) {
     delete group[key];
   }
-  editForm.value.save();
+  editForm.value?.save();
   group.name = name.value;
   group.type = type.value;
   onSave?.();
@@ -50,6 +52,9 @@ function onSaveClick() {
       <Active label="Type">
         <SelectInput v-model="type" :options="typeOptions" />
       </Active>
+      <Passive v-if="shortDescription" label="Description">
+        <span>{{ shortDescription }}</span>
+      </Passive>
       <Active label="Name" :error="nameError">
         <TextInput v-model="name" />
       </Active>