+++ src/features/XIT/ACT/act-registry.ts
@@ -2,19 +2,21 @@ import {
   ActionStep,
   ActionStepExecuteContext,
   ActionStepGenerateContext,
+  MaterialBill,
   MaterialGroupGenerateContext,
 } from '@src/features/XIT/ACT/shared-types';
 
 interface MaterialGroupInfo<TConfig> {
   type: UserData.MaterialGroupType;
+  shortDescription?: string;
   description: (data: UserData.MaterialGroupData, config?: TConfig) => string;
-  editComponent: Component;
+  editComponent?: Component;
   configureComponent?: Component;
   needsConfigure?: (data: UserData.MaterialGroupData) => boolean;
   isValidConfig?: (data: UserData.MaterialGroupData, config: TConfig) => boolean;
   generateMaterialBill: (
     ctx: MaterialGroupGenerateContext<TConfig>,
-  ) => Promise<Record<string, number> | undefined>;
+  ) => Promise<MaterialBill | undefined>;
 }
 
 const materialGroups: MaterialGroupInfo<unknown>[] = [];
@@ -33,6 +35,7 @@ function getMaterialGroupTypes() {
 
 interface ActionInfo<TConfig> {
   type: UserData.ActionType;
+  shortDescription?: string;
   description: (data: UserData.ActionData, config?: TConfig) => string;
   editComponent: Component;
   configureComponent?: Component;