+++ src/features/XIT/ACT/runner/step-machine.ts
@@ -2,8 +2,8 @@ import { act } from '@src/features/XIT/ACT/act-registry';
 import { ActionStep } from '@src/features/XIT/ACT/shared-types';
 import { Logger } from '@src/features/XIT/ACT/runner/logger';
 import { TileAllocator } from '@src/features/XIT/ACT/runner/tile-allocator';
-import { clickElement } from '@src/util';
 import { sleep } from '@src/utils/sleep';
+import { waitActionFeedback } from '@src/infrastructure/prun-ui/utils/action-feedback';
 
 interface StepMachineOptions {
   tile: PrunTile;
@@ -16,7 +16,7 @@ interface StepMachineOptions {
   onActReady: () => void;
 }
 
-const AssertionError = new Error('Assertion failed');
+const CancellationError = new Error('The action was cancelled');
 
 export class StepMachine {
   private next?: ActionStep;
@@ -99,13 +99,19 @@ export class StepMachine {
         },
         waitActionFeedback: async tile => {
           this.options.onStatusChanged('Waiting for action feedback...');
-          const error = await waitActionFeedback(tile);
-          if (error) {
-            log.error(error);
+          const { result, message } = await waitActionFeedback(tile.frame, {
+            autoConfirm: true,
+            dismissSuccess: true,
+          });
+          if (result === 'cancel') {
+            log.cancel('Action Package execution canceled');
+            throw CancellationError;
+          }
+          if (result === 'error') {
+            log.error(message);
             log.error(description ?? info.description(next));
             log.error('Action Package execution failed');
-            this.stop();
-            return;
+            throw CancellationError;
           }
         },
         cacheDescription: () => {
@@ -124,19 +130,18 @@ export class StepMachine {
             log.error(message);
           }
           log.error('Action Package execution failed');
-          this.stop();
-          return;
+          throw CancellationError;
         },
         assert: (condition, message) => {
           if (!condition) {
             log.error(message);
-            throw AssertionError;
+            throw CancellationError;
           }
         },
         requestTile: async command => await this.requestTile(command),
       });
     } catch (e) {
-      if (e !== AssertionError) {
+      if (e !== CancellationError) {
         log.runtimeError(e);
       }
       this.stop();
@@ -153,7 +158,7 @@ export class StepMachine {
     tile = await this.options.tileAllocator.requestTile(command);
     if (tile === undefined) {
       this.log.error(`Failed to open ${command}`);
-      this.stop();
+      throw CancellationError;
     }
     return tile;
   }
@@ -170,43 +175,4 @@ export class StepMachine {
     }
     return this.isRunning;
   }
-}
-
-async function waitActionFeedback(tile: PrunTile) {
-  const overlay = await $(tile.frame, C.ActionFeedback.overlay);
-  await waitActionProgress(overlay);
-  if (overlay.classList.contains(C.ActionConfirmationOverlay.container)) {
-    const confirm = _$$(overlay, C.Button.btn)[1];
-    if (confirm === undefined) {
-      return 'Confirmation overlay is missing confirm button';
-    }
-    await clickElement(confirm);
-    await waitActionProgress(overlay);
-  }
-  if (overlay.classList.contains(C.ActionFeedback.success)) {
-    await clickElement(overlay);
-    return;
-  }
-  if (overlay.classList.contains(C.ActionFeedback.error)) {
-    const message = _$(overlay, C.ActionFeedback.message)?.textContent;
-    const dismiss = _$(overlay, C.ActionFeedback.dismiss)?.textContent;
-    return dismiss ? message?.replace(dismiss, '') : message;
-  }
-
-  return 'Unknown action feedback overlay';
-}
-
-async function waitActionProgress(overlay: HTMLElement) {
-  if (!overlay.classList.contains(C.ActionFeedback.progress)) {
-    return;
-  }
-  await new Promise<void>(resolve => {
-    const mutationObserver = new MutationObserver(() => {
-      if (!overlay.classList.contains(C.ActionFeedback.progress)) {
-        mutationObserver.disconnect();
-        resolve();
-      }
-    });
-    mutationObserver.observe(overlay, { attributes: true });
-  });
 }